refactor: MAT-304 경기 기록 웹소켓 Hook에서 읽기/쓰기 기능 분리
