+++ src/features/matchRecord/index.ts
@@ -1,2 +1,3 @@
-export * from './useMatchRecordWebSocket';
+export * from './useMatchRecordWsMutation';
+export * from './useMatchRecordWsSubscribe';
 export * from './useSyncMatchMemo';
